Disable spell checking keys starting with <leader>s
Problem: It interferes with <leader>s of easy motion's f2 search. And I
also doesn't use this feature at all. We programmers do not need spell
checkers! And the default key for spell checking is already intuitive
and short, so there is no need for extra customization to the original
Vim.